refactor: updated file name / changed file location (#650)
