+++ app/components/assessment/Timer.tsx
@@ -7,7 +7,7 @@ import type {
   SectionInCandidateTest,
   SectionInTest,
 } from "~/interface/Interface"
-import { getTimeLeftInSeconds } from "~/utils/assessment.utils"
+import { getTimeLeftInSeconds } from "~/services/assessment.service"
 
 const TimerComponent = ({
   candidateSection,
