+++ app/routes/assessment/$assessmentId/$sectionId/cooldown.tsx
@@ -7,7 +7,7 @@ import {
   getSectionInCandidateTest,
   getSectionInTest,
   startCandidateSection,
-} from "~/utils/assessment.utils"
+} from "~/services/assessment.service"
 
 export const loader: LoaderFunction = async ({ params, request }) => {
   const section = await getSectionInTest(params.sectionId as string)
