+++ app/routes/assessment/$assessmentId/already-submitted.tsx
@@ -2,7 +2,7 @@ import type { LoaderFunction } from "@remix-run/node"
 import { redirect } from "@remix-run/node"
 
 import AlredySubmitted from "~/components/assessment/AlreadySubmitted"
-import { checkIfTestLinkIsValidAndRedirect } from "~/utils/assessment.utils"
+import { checkIfTestLinkIsValidAndRedirect } from "~/services/assessment.service"
 
 export const loader: LoaderFunction = async ({ params, request }) => {
   const candidateNextRoute = await checkIfTestLinkIsValidAndRedirect(
