+++ app/routes/assessment/$assessmentId/instructions.tsx
@@ -13,7 +13,7 @@ import {
   startCandidateSection,
   startTest,
   updateNextStep,
-} from "~/utils/assessment.utils"
+} from "~/services/assessment.service"
 
 export const loader: LoaderFunction = async ({ params, request }) => {
   const candidateNextRoute = await checkIfTestLinkIsValidAndRedirect(
