+++ app/routes/assessment/$assessmentId/register.tsx
@@ -7,7 +7,7 @@ import {
   getCandidateIDFromAssessmentId,
   updateCandidateDetail,
   updateNextStep,
-} from "~/utils/assessment.utils"
+} from "~/services/assessment.service"
 
 export const loader: LoaderFunction = async ({ params, request }) => {
   const candidateNextRoute = await checkIfTestLinkIsValidAndRedirect(
